create a python package for string method
